PULLA TUTTO CHE COSI E' OK (a parte forse il TopCoach)

+++ php/compravendi-exe.php
@@ -3,8 +3,6 @@
 session_start();
 include_once("connessione.php");
 
-
-
 $nick=$_SESSION['nick'];
 
 $vecchioPort=$_POST['vendiPortieri'];
@@ -19,16 +17,6 @@ $nuoviAtt =$_POST['compraAttaccanti'];
 
 $mod="NO";
 
-//CANCELLO FORMAZIONI SE CAMBIO QUALCOSA
-if($vecchioPort!=0 || $vecchioDif!=0 ||$vecchioCent!=0 ||$vecchioAtt!=0||
-		$nuoviPort!=0 || $nuoviDif!=0 ||$nuoviCent!=0 ||$nuoviAtt!=0){
-	$query="DELETE FROM formazione
-					WHERE Squadra='$nomeSq[0]'";
-	$cid->query($query);
-	$mod="SI";
-
-}
-
  // Salvo in una variabile il nome della squadra loggata
 $sql="SELECT nomeSq FROM squadra JOIN utente ON Mail=Utente WHERE Nickname='".$nick."'";
 $squadra=$cid->query($sql) or die("<p>Impossibile eseguire query.</p>"
@@ -37,13 +25,16 @@ $squadra=$cid->query($sql) or die("<p>Impossibile eseguire query.</p>"
 $nomeSq=$squadra->fetch_row();
 
 //CANCELLO FORMAZIONI SE CAMBIO QUALCOSA
-if($vecchioPort!=0 || $vecchioDif!=0 ||$vecchioCent!=0 ||$vecchioAtt!=0){
+if($vecchioPort!=0 || $vecchioDif!=0 ||$vecchioCent!=0 ||$vecchioAtt!=0||
+		$nuoviPort!=0 || $nuoviDif!=0 ||$nuoviCent!=0 ||$nuoviAtt!=0){
 	$query="DELETE FROM formazione
 					WHERE Squadra='$nomeSq[0]'";
 	$cid->query($query);
+	// Attributo che verrà passato in url per capire se ci sono state modifiche
+	$mod="SI";
 }
 
-//calcolo il numero portieri (e poi tutti i ruoli) che ho in rosa
+// calcolo il numero portieri (e poi tutti i ruoli) che ho in rosa
 $sql="SELECT COUNT(*)
 				FROM possiede JOIN giocatore ON (Giocatore=Cognome)
 				WHERE Ruolo='P' AND SquadraGioc='$nomeSq[0]'";
@@ -78,16 +69,17 @@ $sql="SELECT COUNT(*)
 	$nAtt=$res->fetch_row();
 
 
-//conteggio i fantaMilioni
-// guardo i milioni che mi rimangono
+//conteggio i fantaMilioni rimanenti
 $sql="SELECT SUM(Prezzo) FROM giocatore JOIN possiede ON giocatore=cognome
 					WHERE SquadraGioc= (SELECT nomeSq FROM squadra JOIN utente ON Mail=Utente WHERE Nickname='$nick')";
 $fantamilioniSpesi=$cid->query($sql) or die("<p>Impossibile eseguire query.</p>"
 															."<p>codice di errore ".$cid->errno
 															.":".$cid->error."</p>");
 $fantasoldi=$fantamilioniSpesi->fetch_row();
-
+// Aggiorno il valore
 $fantamilioni=300-$fantasoldi[0];
+
+// Calcolo i prezzi dei portieri che voglio vendere
 $soldiVenditaPortieri=0;
 foreach($vecchioPort as $VP) {
 	$sql="SELECT Prezzo FROM Giocatore
@@ -98,6 +90,7 @@ foreach($vecchioPort as $VP) {
 	$soldi=$denaro->fetch_row();
 	$soldiVenditaPortieri=$soldiVenditaPortieri+$soldi[0];
 }
+// Calcolo i prezzi dei portieri che voglio comprare
 $soldiAquistoPortieri=0;
 foreach($nuoviPort as $NP) {
 	$sql="SELECT Prezzo FROM Giocatore
@@ -108,6 +101,7 @@ foreach($nuoviPort as $NP) {
 	$soldi=$denaro->fetch_row();
 	$soldiAquistoPortieri=$soldiAquistoPortieri+$soldi[0];
 }
+// Calcolo i prezzi dei difensori che voglio vendere
 $soldiVenditaDifensori=0;
 foreach($vecchioDif as $VD) {
 	$sql="SELECT Prezzo FROM Giocatore
@@ -118,6 +112,7 @@ foreach($vecchioDif as $VD) {
 	$soldi=$denaro->fetch_row();
 	$soldiVenditaDifensori=$soldiVenditaDifensori+$soldi[0];
 }
+// Calcolo i prezzi dei difensori che voglio comprare
 $soldiAquistoDifensori=0;
 foreach($nuoviDif as $ND) {
 	$sql="SELECT Prezzo FROM Giocatore
@@ -128,6 +123,7 @@ foreach($nuoviDif as $ND) {
 	$soldi=$denaro->fetch_row();
 	$soldiAquistoDifensori=$soldiAquistoDifensori+$soldi[0];
 }
+// Calcolo i prezzi dei centrocampisti che voglio vendere
 $soldiVenditaCentrocampisti=0;
 foreach($vecchioCent as $VC) {
 	$sql="SELECT Prezzo FROM Giocatore
@@ -138,6 +134,7 @@ foreach($vecchioCent as $VC) {
 	$soldi=$denaro->fetch_row();
 	$soldiVenditaCentrocampisti=$soldiVenditaCentrocampisti+$soldi[0];
 }
+// Calcolo i prezzi dei centrocampisti che voglio comprare
 $soldiAquistoCentrocampisti=0;
 foreach($nuoviCent as $NC) {
 	$sql="SELECT Prezzo FROM Giocatore
@@ -148,6 +145,7 @@ foreach($nuoviCent as $NC) {
 	$soldi=$denaro->fetch_row();
 	$soldiAquistoCentrocampisti=$soldiAquistoCentrocampisti+$soldi[0];
 }
+// Calcolo i prezzi degli attaccanti che voglio vendere
 $soldiVenditaAttaccanti=0;
 foreach($vecchioAtt as $VA) {
 	$sql="SELECT Prezzo FROM Giocatore
@@ -158,6 +156,7 @@ foreach($vecchioAtt as $VA) {
 	$soldi=$denaro->fetch_row();
 	$soldiVenditaAttaccanti=$soldiVenditaAttaccanti+$soldi[0];
 }
+// Calcolo i prezzi degli attaccanti che voglio comprare
 $soldiAquistoAttaccanti=0;
 foreach($nuoviAtt as $NA) {
 	$sql="SELECT Prezzo FROM Giocatore
@@ -170,42 +169,47 @@ foreach($nuoviAtt as $NA) {
 
 }
 
+	// Se non si sfora il budget
+	if((300
+			-$fantasoldi[0]
+			-$soldiAquistoPortieri
+			-$soldiAquistoDifensori
+			-$soldiAquistoCentrocampisti
+			-$soldiAquistoAttaccanti
+			+$soldiVenditaPortieri
+			+$soldiVenditaDifensori
+			+$soldiVenditaCentrocampisti
+			+$soldiVenditaAttaccanti)>0) {
 
-					if((300
-							-$fantasoldi[0]
-							-$soldiAquistoPortieri
-							-$soldiAquistoDifensori
-							-$soldiAquistoCentrocampisti
-							-$soldiAquistoAttaccanti
-							+$soldiVenditaPortieri
-							+$soldiVenditaDifensori
-							+$soldiVenditaCentrocampisti
-							+$soldiVenditaAttaccanti)>0){
-
+					// Se non si sfora il numero di giocatori
 					if(($nPor[0]+$nDif[0]+$nAtt[0]+$nCen[0]
 							- count($vecchioPort) + count($nuoviPort)
 							- count($vecchioDif) + count($nuoviDif)
 							- count($vecchioCent) + count($nuoviCent)
 							- count($vecchioAtt) + count($nuoviAtt))<12){
-					//se rispetta le condizioni di numero c'è un ingresso da parte di ogni ruolo
-					//portieri max 2
-								if(($nPor[0] - count($vecchioPort) + count($nuoviPort))<3){
 
+
+							/*     ----- CONTROLLO REGOLARITA' COMPRAVENDITA, MODIFICHE E MESSAGGI DI ERRORE ----- */
+								//se rispetta le condizioni di numero di giocatori per ruolo effettuo l'aggiornamento
+								if(($nPor[0] - count($vecchioPort) + count($nuoviPort))<3){
+										// Cancella vecchi portieri
 										foreach($vecchioPort as $VP) {
 											$query="DELETE FROM possiede
 										          WHERE Giocatore='$VP' AND SquadraGioc='$nomeSq[0]'";
 											$cid->query($query);
 										}
 											$fantamilioni=$fantamilioni+$soldiVenditaPortieri;
-
+											// Compra nuovi portieri
 										foreach($nuoviPort as $NP) {
 											$query="INSERT INTO possiede (Giocatore,	SquadraGioc)
 															VALUES ('$NP','$nomeSq[0]')";
 											$cid->query($query);
 										}
+											// Agiorno il budget e messaggio che si passa in url
 											$fantamilioni=$fantamilioni-$soldiAquistoPortieri;
 											$msgPortieri="OK";
 								}
+								// messaggio di errore
 								else {
 										$msgPortieri="over_P";
 								}
@@ -288,15 +292,12 @@ foreach($nuoviAtt as $NA) {
 					}
 	$msgFineSoldi="OK";
 	}
-
+	// Se si ha superato il budget consentito
 	else {
 		$msgFineSoldi="over_B";
 	}
 
-
-
-
-
+// Mando nell'url tutte le sigle per la gestione dei messaggi di errore
 header("Location:../index.php?op=fantamercato&mod=$mod&msgP=$msgPortieri&msgD=$msgDifensori&msgC=$msgCentrocampisti&msgA=$msgAttaccanti&msgN=$msgEccesso&msgB=$msgFineSoldi");
